requirements.txt
stegano
tkinter
pillow
